Move JointStateListener to library
Move code to a library and add an separate executable with the main
function.
