+++ src/multi_robot_state_publisher_node.cpp
@@ -1,0 +1,98 @@
+// Copyright 2021 Ryan Sinnet
+
+/*********************************************************************
+ * Software License Agreement (BSD License)
+ *
+ *  Copyright (c) 2008, Willow Garage, Inc.
+ *  All rights reserved.
+ *
+ *  Redistribution and use in source and binary forms, with or without
+ *  modification, are permitted provided that the following conditions
+ *  are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *   * Redistributions in binary form must reproduce the above
+ *     copyright notice, this list of conditions and the following
+ *     disclaimer in the documentation and/or other materials provided
+ *     with the distribution.
+ *   * Neither the name of the Willow Garage nor the names of its
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ *  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ *  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ *  FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ *  COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ *  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ *  BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ *  LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ *  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ *  LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ *  POSSIBILITY OF SUCH DAMAGE.
+ *********************************************************************/
+
+/* Author: Wim Meeussen */
+
+#include <map>
+#include <string>
+
+#include <kdl/frames_io.hpp>
+#include <kdl_parser/kdl_parser.hpp>
+#include <ros/ros.h>
+#include <tf2_kdl/tf2_kdl.h>
+#include <urdf/model.h>
+
+#include "multi_robot_state_publisher/joint_state_listener.h"
+#include "multi_robot_state_publisher/robot_state_publisher.h"
+
+using multi_robot_state_publisher::JointStateListener;
+using multi_robot_state_publisher::MimicMap;
+
+int main(int argc, char** argv)
+{
+  ros::init(argc, argv, "multi_robot_state_publisher");
+  ros::NodeHandle node;
+
+  ///////////////////////////////////////// begin deprecation warning
+  std::string exe_name = argv[0];
+  std::size_t slash = exe_name.find_last_of("/");
+  if (slash != std::string::npos)
+  {
+    exe_name = exe_name.substr(slash + 1);
+  }
+  if (exe_name == "state_publisher")
+  {
+    ROS_WARN("The 'state_publisher' executable is deprecated. Please use 'multi_robot_state_publisher' instead");
+  }
+  ///////////////////////////////////////// end deprecation warning
+
+  // gets the location of the robot description on the parameter server
+  urdf::Model model;
+  if (!model.initParam("robot_description"))
+    return -1;
+
+  KDL::Tree tree;
+  if (!kdl_parser::treeFromUrdfModel(model, tree))
+  {
+    ROS_ERROR("Failed to extract kdl tree from xml robot description");
+    return -1;
+  }
+
+  MimicMap mimic;
+
+  for (std::map<std::string, urdf::JointSharedPtr>::iterator i = model.joints_.begin(); i != model.joints_.end(); i++)
+  {
+    if (i->second->mimic)
+    {
+      mimic.insert(make_pair(i->first, i->second->mimic));
+    }
+  }
+
+  JointStateListener state_publisher(tree, mimic, model);
+  ros::spin();
+
+  return 0;
+}
